requirements.txt
google-adk==1.13.0
google-cloud-bigquery==3.36.0
pdfplumber==0.11.7
pypdf2==3.0.1
python-docx==1.2.0
python-dotenv==1.2.1
fastapi==0.120.0
uvicorn==0.38.0
pydantic==2.11.7

python-multipart
PyPDF2

